server/requirements.txt
torch
flask
pillow